+++ src/main/java/dev/liqw/persistentchat/client/mixin/ChatComponentMixin.java
@@ -4,7 +4,10 @@ import dev.liqw.persistentchat.client.utils.StateManager;
 import dev.liqw.persistentchat.client.utils.MessageTimestampRegistry;
 import dev.liqw.persistentchat.utils.GuiMessageAccessor;
 import net.minecraft.client.gui.components.ChatComponent;
+//? >=26 {
 import net.minecraft.client.multiplayer.chat.GuiMessage;
+//? } else
+//import net.minecraft.client.GuiMessage;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;