+++ src/main/java/dev/liqw/persistentchat/client/mixin/GuiMessageMixin.java
@@ -1,7 +1,10 @@
 package dev.liqw.persistentchat.client.mixin;
 
 import dev.liqw.persistentchat.utils.GuiMessageAccessor;
+//? >=26 {
 import net.minecraft.client.multiplayer.chat.GuiMessage;
+ //? } else
+//import net.minecraft.client.GuiMessage;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;